machine: lazy load tpi/python_terraform only when needed
