+++ dvc/machine/backend/terraform.py
@@ -1,13 +1,10 @@
 import logging
 import os
-import tempfile
 from contextlib import contextmanager
 from typing import TYPE_CHECKING, Iterator, Optional
 
 from funcy import first
-from python_terraform import Terraform
 
-from dvc.exceptions import DvcException
 from dvc.fs.ssh import SSHFileSystem
 
 from .base import BaseMachineBackend
@@ -18,51 +15,10 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class TerraformError(DvcException):
-    pass
-
-
-class DvcTerraform(Terraform):
-    def cmd(self, *args, **kwargs):
-        logger.debug(" ".join(self.generate_cmd_string(*args, **kwargs)))
-        kwargs["capture_output"] = False
-        ret, _stdout, _stderr = super().cmd(*args, **kwargs)
-        if ret != 0:
-            raise TerraformError("Cmd 'terraform {cmd}' failed")
-
-    def iter_instances(self, name: str):
-        """Iterate over active iterative_machine instances."""
-        self.read_state_file()
-        resources = getattr(self.tfstate, "resources", [])
-        for resource in resources:
-            if (
-                resource.get("type") == "iterative_machine"
-                and resource.get("name") == name
-            ):
-                yield from (
-                    instance.get("attributes", {})
-                    for instance in resource.get("instances", [])
-                )
-
-    @staticmethod
-    @contextmanager
-    def pemfile(resource: dict):
-        """Generate a temporary SSH key (PEM) file for the specified resource.
-
-        The path to the file will be yielded, and the file will be removed
-        upon exiting the yielded context.
-        """
-        with tempfile.NamedTemporaryFile(delete=False) as fobj:
-            fobj.write(resource["ssh_private"].encode("ascii"))
-        try:
-            yield fobj.name
-        finally:
-            os.unlink(fobj.name)
-
-
 class TerraformBackend(BaseMachineBackend):
     @contextmanager
     def make_tf(self, name: str):
+        from dvc.tpi import DvcTerraform, TerraformError
         from dvc.utils.fs import makedirs
 
         try:
@@ -110,6 +66,8 @@ class TerraformBackend(BaseMachineBackend):
         raise NotImplementedError
 
     def _default_resource(self, name):
+        from dvc.tpi import TerraformError
+
         resource = first(self.instances(name))
         if not resource:
             raise TerraformError(f"No active '{name}' instances")
@@ -119,6 +77,8 @@ class TerraformBackend(BaseMachineBackend):
     def get_sshfs(
         self, name: Optional[str] = None, **config
     ) -> Iterator["SSHFileSystem"]:
+        from dvc.tpi import DvcTerraform
+
         resource = self._default_resource(name)
         with DvcTerraform.pemfile(resource) as pem:
             fs = SSHFileSystem(
@@ -129,6 +89,8 @@ class TerraformBackend(BaseMachineBackend):
             yield fs
 
     def run_shell(self, name: Optional[str] = None, **config):
+        from dvc.tpi import DvcTerraform
+
         resource = self._default_resource(name)
         with DvcTerraform.pemfile(resource) as pem:
             self._shell(
